Activity Selection-Greedy Algorithm
The greedy choice is to always pick the next activity whose finish time
is least among the remaining activities and the start time is more than
or equal to the finish time of previously selected activity. We can sort
the activities according to their finishing time so that we always
consider the next activity as minimum finishing time activity.